Moved evm related tests out into their own files
